Organize MainScreen (MainActivity) with the User TrackMaps and the bottom navigation component.
