+++ app/src/main/AndroidManifest.xml
@@ -24,31 +24,25 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
-        <activity android:name=".features.create.CreateActivity"
-            android:label="Create a Trackmap">
-            <!--<intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>-->
-        </activity>
 
         <meta-data
             android:name="com.google.android.geo.API_KEY"
             android:value="@string/google_maps_key" />
 
-        <activity
-            android:name=".features.entries.CurrentTrackMapsActivity"
-            android:theme="@style/Theme.AppCompat.Light.NoActionBar" />
-        <activity
-            android:name=".features.trackmap.MainActivity"
-            android:label="@string/app_name">
+        <activity android:name=".features.entries.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-    </application>
 
+        <activity
+            android:name=".features.trackmap.TrackMapActivity"
+            android:label="@string/app_name" />
+
+        <activity
+            android:name=".features.create.CreateActivity"
+            android:label="Create a Trackmap" />
+    </application>
 </manifest>
