+++ app/src/main/java/com/handysparksoft/trackmap/core/di/AppComponent.kt
@@ -1,11 +1,10 @@
 package com.handysparksoft.trackmap.core.di
 
 import android.app.Application
-import com.handysparksoft.trackmap.core.platform.UserHandler
 import com.handysparksoft.trackmap.features.create.CreateViewModelFactory
-import com.handysparksoft.trackmap.features.entries.CurrentTrackMapsViewModelFactory
-import com.handysparksoft.trackmap.features.trackmap.MainActivity
-import com.handysparksoft.trackmap.features.trackmap.MainViewModelFactory
+import com.handysparksoft.trackmap.features.entries.MainViewModelFactory
+import com.handysparksoft.trackmap.features.trackmap.TrackMapActivity
+import com.handysparksoft.trackmap.features.trackmap.TrackMapViewModelFactory
 import dagger.BindsInstance
 import dagger.Component
 import javax.inject.Singleton
@@ -17,13 +16,13 @@ interface AppComponent {
     // Exposed Graph components
     val mainViewModelFactory: MainViewModelFactory
     val createViewModelFactory: CreateViewModelFactory
-    val currentViewModelFactory: CurrentTrackMapsViewModelFactory
+    val trackMapViewModelFactory: TrackMapViewModelFactory
 
     // Field injections
-    fun inject(mainActivity: MainActivity)
+    fun inject(trackMapActivity: TrackMapActivity)
 
     @Component.Factory
     interface Factory {
-        fun create(@BindsInstance app: Application) : AppComponent
+        fun create(@BindsInstance app: Application): AppComponent
     }
 }
