+++ app/src/main/res/drawable/button_primary_medium_shape.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <solid android:color="@color/colorAccent"/>
+    <corners android:radius="20dp"/>
+</shape>
