+++ app/src/main/res/values/colors.xml
@@ -3,4 +3,75 @@
     <color name="colorPrimary">#008577</color>
     <color name="colorPrimaryDark">#00574B</color>
     <color name="colorAccent">#D81B60</color>
+
+    <color name="color_background">#FFFFFF</color>
+    <color name="color_background_secondary">#F9FAFA</color>
+    <color name="color_background_tertiary">#F4F4F5</color>
+    <color name="color_beta">#F4164D</color>
+    <color name="color_border">#BCC9D1</color>
+    <color name="color_border_soft">#DBE2E6</color>
+    <color name="color_border_strong">#1D2124</color>
+    <color name="color_caution">#FFA000</color>
+    <color name="color_control_border">#BCC9D1</color>
+    <color name="color_control_border_secondary">#1D2124</color>
+    <color name="color_control_border_tertiary">#FFFFFF</color>
+    <color name="color_control_fill">#FFFFFF</color>
+    <color name="color_control_fill_quaternary">#1D2124</color>
+    <color name="color_control_fill_secondary">#F4F4F5</color>
+    <color name="color_control_fill_tertiary">#DBE2E6</color>
+    <color name="color_disabled">#DBE2E6</color>
+    <color name="color_disabled_soft">#E9EEF0</color>
+    <color name="color_disabled_strong">#BCC9D1</color>
+    <color name="color_error">#F4164D</color>
+    <color name="color_error_soft">#F8DDDF</color>
+    <color name="color_focussed">#009090</color>
+    <color name="color_go">#4DB86B</color>
+    <color name="color_highlighted">#C6F16D</color>
+    <color name="color_hover">#F4F4F5</color>
+    <color name="color_hover_quaternary">#414D5D</color>
+    <color name="color_hover_secondary">#E9EEF0</color>
+    <color name="color_hover_tertiary">#BCC9D1</color>
+    <color name="color_info">#68C3FC</color>
+    <color name="color_info_soft">#E3EDFE</color>
+    <color name="color_link">#009090</color>
+    <color name="color_link_hover">#007575</color>
+    <color name="color_link_visited">#728895</color>
+    <color name="color_new">#F4164D</color>
+    <color name="color_notification">#F4164D</color>
+    <color name="color_premium">#009090</color>
+    <color name="color_premium_quaternary">#CFE3E3</color>
+    <color name="color_premium_secondary">#007875</color>
+    <color name="color_premium_tertiary">#E4F0F0</color>
+    <color name="color_pricing">#9C195B</color>
+    <color name="color_primary_button_hover">#B5DA67</color>
+    <color name="color_pro_business">#73418D</color>
+    <color name="color_pro_business_quaternary">#E1CFEE</color>
+    <color name="color_pro_business_secondary">#663082</color>
+    <color name="color_pro_business_tertiary">#F2F5F9</color>
+    <color name="color_pro_jobs">#9C195B</color>
+    <color name="color_pro_jobs_quaternary">#E5BCD0</color>
+    <color name="color_pro_jobs_secondary">#940F62</color>
+    <color name="color_pro_jobs_tertiary">#F2F5F9</color>
+    <color name="color_selected">#1D2124</color>
+    <color name="color_selected_secondary">#009090</color>
+    <color name="color_selected_tertiary">#BFDCDC</color>
+    <color name="color_skeleton_primary">#E9EEF0</color>
+    <color name="color_skeleton_primary_animation">#DBE2E6</color>
+    <color name="color_skeleton_secondary">#BCC9D1</color>
+    <color name="color_skeleton_secondary_animation">#A4B2BA</color>
+    <color name="color_stop">#F4164D</color>
+    <color name="color_success">#4DB86B</color>
+    <color name="color_success_soft">#CFF1D8</color>
+    <color name="color_text">#1D2124</color>
+    <color name="color_text_invert">#FFFFFF</color>
+    <color name="color_text_on_dark">#FFFFFF</color>
+    <color name="color_text_on_light">#1D2124</color>
+    <color name="color_text_secondary">#5E737D</color>
+    <color name="color_text_secondary_on_dark">#8A9DA6</color>
+    <color name="color_text_secondary_on_light">#5E737D</color>
+    <color name="color_text_tertiary">#8A9DA6</color>
+    <color name="color_text_tertiary_on_dark">#5E737D</color>
+    <color name="color_text_tertiary_on_light">#8A9DA6</color>
+    <color name="color_unseen">#F4164D</color>
+    <color name="color_user_roles">#7C8A9C</color>
 </resources>
