+++ app/src/main/res/values/styles.xml
@@ -6,6 +6,43 @@
         <item name="colorPrimary">@color/colorPrimary</item>
         <item name="colorPrimaryDark">@color/colorPrimaryDark</item>
         <item name="colorAccent">@color/colorAccent</item>
+
+        <item name="android:windowBackground">@color/color_background_secondary</item>
+    </style>
+
+    <!-- Button Styles -->
+    <style name="Primary" parent="Widget.MaterialComponents.Button.UnelevatedButton">
+        <item name="android:textAllCaps">false</item>
+        <item name="backgroundTint">@color/colorPrimary</item>
+    </style>
+
+    <style name="Primary.Small">
+        <item name="android:layout_height">30dp</item>
+        <item name="android:padding">0dp</item>
+        <item name="android:textSize">10dp</item>
+        <item name="cornerRadius">15dp</item>
+    </style>
+
+    <style name="Primary.Medium">
+        <item name="android:layout_height">40dp</item>
+        <item name="cornerRadius">20dp</item>
+    </style>
+
+    <style name="Secondary" parent="Widget.MaterialComponents.Button.OutlinedButton">
+        <item name="android:textAllCaps">false</item>
+        <item name="strokeColor">@color/colorPrimary</item>
+    </style>
+
+    <style name="Secondary.Small">
+        <item name="android:layout_height">30dp</item>
+        <item name="android:padding">0dp</item>
+        <item name="android:textSize">10dp</item>
+        <item name="cornerRadius">15dp</item>
+    </style>
+
+    <style name="Secondary.Medium">
+        <item name="android:layout_height">40dp</item>
+        <item name="cornerRadius">20dp</item>
     </style>
 
 </resources>
